Rename `timeouts` to `cancellation` module
