+++ src/prefect/_internal/concurrency/threads.py
@@ -12,7 +12,7 @@ from typing import List, Optional
 from prefect._internal.concurrency.calls import Call, Portal
 from prefect._internal.concurrency.event_loop import get_running_loop
 from prefect._internal.concurrency.primitives import Event
-from prefect._internal.concurrency.timeouts import CancelledError
+from prefect._internal.concurrency.cancellation import CancelledError
 from prefect._internal.concurrency import logger
