+++ tests/_internal/concurrency/test_api.py
@@ -8,7 +8,7 @@ import pytest
 
 from prefect._internal.concurrency.api import create_call, from_async, from_sync
 from prefect._internal.concurrency.threads import wait_for_global_loop_exit
-from prefect._internal.concurrency.timeouts import CancelledError
+from prefect._internal.concurrency.cancellation import CancelledError
 
 
 def identity(x):
