+++ tests/_internal/concurrency/test_waiters.py
@@ -4,7 +4,7 @@ import pytest
 
 from prefect._internal.concurrency.calls import Call
 from prefect._internal.concurrency.threads import WorkerThread
-from prefect._internal.concurrency.timeouts import CancelledError
+from prefect._internal.concurrency.cancellation import CancelledError
 from prefect._internal.concurrency.waiters import AsyncWaiter, SyncWaiter
